final-day commit

+++ admin/script/blog_delete.php
@@ -8,24 +8,41 @@ $id = $_REQUEST['id'];
 
 $sql->autocommit(FALSE); 
 
-$query_tag_delete = "DELETE FROM `article_tag_map` WHERE `post_id` = $id";
-$result = $sql->query($query_tag_delete);
+$query = "SELECT * FROM `article` WHERE `post_id` = '$id'";
 
-$query_article_delete = "DELETE FROM `article` WHERE `post_id` = $id";
-$result = $sql->query($query_article_delete);
+$result = $sql->query( $query );
 
-if($result)
+while ( $row = $result->fetch_assoc() ) 
 {
-	echo 'success';	
-	$sql->commit();
-	$sql->close();
-}	
-else
-{ 
-    	echo 'error';
-	$sql->rollback();
-  	$sql->close();
+  	$file = $row['img_url'];
+} 
+
+if(unlink($file))
+{
+
+	$query_tag_delete = "DELETE FROM `article_tag_map` WHERE `post_id` = $id";
+
+	$result = $sql->query($query_tag_delete);
+
+	$query_article_delete = "DELETE FROM `article` WHERE `post_id` = $id";
+
+	$result = $sql->query($query_article_delete);
+
+	if($result)
+	{
+		echo 'success';	
+		$sql->commit();
+		$sql->close();
+	}	
+	else
+	{ 
+	    echo 'error';
+		$sql->rollback();
+	  	$sql->close();
+	}
+
 }
 
-?>
 
+
+?>
